multimodal-search-tirg/requirements.txt
click==7.1.2
jina[python-magic]==1.1.0
matplotlib==3.3.3
pillow==8.1.1